renamed the file correctly and set up validation
